[DEV-243] Waiting 서버 유즈케이스 작성

+++ src/main/kotlin/com/tiketeer/TiketeerWaiting/domain/waiting/usecase/dto/GetRankAndTokenResultDto.kt
@@ -1,6 +1,6 @@
 package com.tiketeer.TiketeerWaiting.domain.waiting.usecase.dto
 
 data class GetRankAndTokenResultDto(
-    val rank: Int,
+    val rank: Long,
     val token: String? = null,
 )
